requirements.txt
asgiref==3.4.1
certifi==2021.10.8
click==8.0.3
Django==3.2.9
django-debug-toolbar==3.2.2
django-environ==0.8.1
Flask==2.0.2
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
psycopg2==2.8.4
pytz==2021.3
sqlparse==0.4.2
Werkzeug==2.0.2
whitenoise==5.3.0
